backend/requirements.txt
fastapi>=0.104.1
uvicorn>=0.24.0
python-multipart>=0.0.6
pytesseract>=0.3.10
pdf2image>=1.16.3
Pillow>=10.0.0
python-dotenv>=1.0.0
pydantic>=2.4.2 